New SB_Excel y SB_Word; Added funcionalty of route variable
All functions that interact with Excel or Word have been moved to SB_Excel and SB_Word.
A new "route" variable functionality has been created that removes bookmarks from the template when the value is "false." It is handled like any other variable at the tool or datamaster level.
The functionality to generate an error if any of the variable IDs contain denormalized characters has been added: it displays the list of variables with denormalized characters in the console and raises an exception.